+++ admin/write-post.php
@@ -2,7 +2,7 @@
 include 'common.php';
 include 'header.php';
 include 'menu.php';
-\Widget\Contents\Post\Edit::alloc()->to($post);
+\Widget\Contents\Post\Edit::alloc()->prepare()->to($post);
 ?>
 <div class="main">
     <div class="body container">
